Removing lists of collections in feed.
