+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/detail/info/EventInfo.kt
@@ -13,6 +13,7 @@ import androidx.compose.ui.unit.dp
 import com.jordangellatly.hermosahappyhour.model.Restaurant
 import com.jordangellatly.hermosahappyhour.model.SpecialsCollection
 import com.jordangellatly.hermosahappyhour.model.tower12RestaurantData
+import com.jordangellatly.hermosahappyhour.ui.detail.FeaturedSpecialsCollection
 import com.jordangellatly.hermosahappyhour.ui.detail.TodaysEventItem
 import com.jordangellatly.hermosahappyhour.ui.home.*
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
